Python script for coding exercise
This script (FindRestaurants.py) reads a the rest_hours.json file (schedules for various restaurants), prompts the user for a date and time, and displays restaurants that are open at that time. The data file must be in the same folder as the script at run time. It has options for running unit tests or for dumping the parsed restaurant schedules.